frontend: Make store types more closely reflect reality.
