+++ packages/frontend-web/src/app/components/CommentList/index.ts
@@ -20,8 +20,8 @@ import { isNaN, isNumber } from 'lodash';
 import { connect } from 'react-redux';
 import { ICommentModel, ICommentScoredModel } from '../../../models';
 import { ICommentAction } from '../../../types';
+import { IAppDispatch } from '../../appstate';
 import { DATE_FORMAT_HM, DATE_FORMAT_MDY } from '../../config';
-import { IAppDispatch } from '../../stores';
 import { getComment, loadComment } from '../../stores/comments';
 import { getTags } from '../../stores/tags';
 import { getTopScoreForComment, getTopSummaryScoreForComment, loadTopScore, loadTopSummaryScore } from '../../stores/topScores';
