+++ packages/frontend-web/src/app/scenes/Comments/components/ModeratedComments/store/checkedSelection.ts
@@ -15,7 +15,7 @@ limitations under the License.
 */
 
 import { Action, Reducer } from 'redux-actions';
-import { IAppStateRecord } from '../../../../../stores';
+import { IAppStateRecord } from '../../../../../appstate';
 import { ICheckedSelectionPayloads, ICheckedSelectionStateRecord, makeCheckedSelectionStore } from '../../../../../util';
 import { DATA_PREFIX } from './reduxPrefix';
