+++ packages/frontend-web/src/app/scenes/Comments/components/NewComments/index.ts
@@ -20,7 +20,7 @@ import { compose } from 'redux';
 import { createStructuredSelector } from 'reselect';
 
 import { ICommentAction } from '../../../../../types';
-import { IAppDispatch, IAppStateRecord } from '../../../../stores';
+import { IAppDispatch, IAppStateRecord } from '../../../../appstate';
 import { getArticle } from '../../../../stores/articles';
 import {
   approveComments,
