+++ packages/frontend-web/src/app/scenes/Search/components/index.ts
@@ -20,7 +20,7 @@ import { compose } from 'redux';
 import { createStructuredSelector } from 'reselect';
 
 import { ICommentAction } from '../../../../types';
-import { IAppDispatch, IAppStateRecord } from '../../../stores';
+import { IAppDispatch, IAppStateRecord } from '../../../appstate';
 import {
   approveComments,
   deferComments,
