+++ packages/frontend-web/src/app/scenes/Search/store/searchResults.ts
@@ -16,7 +16,7 @@ limitations under the License.
 
 import { List } from 'immutable';
 import { Action, createAction, handleActions } from 'redux-actions';
-import { IAppStateRecord } from '../../../stores';
+import { IAppStateRecord } from '../../../appstate';
 import { DATA_PREFIX } from './reduxPrefix';
 
 export const loadAllCommentIdsStart: () => Action<void> = createAction(
