+++ packages/frontend-web/src/app/stores/appstate.ts
@@ -14,7 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import { Action, Dispatch } from 'redux';
 import { TypedRecord } from 'typed-immutable-record';
 
 import { IArticlesState } from './articles';
@@ -30,7 +29,7 @@ import { ITextSizesStateRecord } from './textSizes';
 import { IState as ITopScoresState, ISummaryState as ITopSummaryScoresState } from './topScores';
 import { IUsersState } from './users';
 
-export interface IAppState {
+export interface IGlobalState {
   categories: ICategoriesState;
   articles: IArticlesState;
   comments: ICommentsState;
@@ -46,11 +45,4 @@ export interface IAppState {
   topSummaryScores: ITopSummaryScoresState;
 }
 
-export interface IAppStateRecord extends TypedRecord<IAppStateRecord>, IAppState {}
-
-export type IThunkAction<R> = (dispatch: Dispatch<IAppStateRecord>, getState: () => IAppStateRecord) => R;
-export type IAction<T> = IThunkAction<T> | Action;
-
-export interface IAppDispatch {
-  <R>(action: IAction<R>): R;
-}
+export interface IGlobalStateRecord extends TypedRecord<IGlobalStateRecord>, IGlobalState {}
