+++ packages/frontend-web/src/app/stores/categories.ts
@@ -17,7 +17,7 @@ limitations under the License.
 import { Action, createAction, handleActions } from 'redux-actions';
 
 import { ICategoryModel, ModelId } from '../../models';
-import { IAppStateRecord } from './appstate';
+import { IAppStateRecord } from '../appstate';
 
 const STATE_ROOT = ['global', 'categories'];
