+++ packages/frontend-web/src/app/stores/counts.ts
@@ -16,7 +16,7 @@ limitations under the License.
 
 import { Action, createAction, handleActions } from 'redux-actions';
 import { makeTypedFactory, TypedRecord } from 'typed-immutable-record';
-import { IAppStateRecord } from './appstate';
+import { IAppStateRecord } from '../appstate';
 
 const STATE_ROOT = ['global', 'counts'];
 const ASSIGNMENTS = [...STATE_ROOT, 'assignments'];
