+++ packages/frontend-web/src/app/stores/focus.ts
@@ -18,7 +18,7 @@ import { List } from 'immutable';
 import { Action, createAction, handleActions } from 'redux-actions';
 import { makeTypedFactory, TypedRecord } from 'typed-immutable-record';
 
-import { IAppStateRecord } from './appstate';
+import { IAppStateRecord } from '../appstate';
 
 export const focusedElement = createAction<string | null>('root/FOCUSED_ELEMENT');
 export const unfocusedElement: () => Action<void> = createAction('root/UNFOCUSED_ELEMENT');
