+++ packages/frontend-web/src/app/stores/preselects.ts
@@ -17,7 +17,7 @@ limitations under the License.
 import { List } from 'immutable';
 import { Action, createAction, handleActions } from 'redux-actions';
 import { IPreselectModel } from '../../models';
-import { IAppStateRecord } from './appstate';
+import { IAppStateRecord } from '../appstate';
 
 const STATE_ROOT = ['global', 'preselects'];
